Add some basic rendering tests.
The Rendering tests simply test that the component renders differently
when it receives different props. It tests every prop that exists in
the current iteration.

There are some changes that have been made to the JS to make tests
easier to work with. These have a small effect on the surface API:

 - If you omit the `color` prop, then `fill` attr is omitted
 - If you omit the `background` prop, then `style` attr is omitted
 - If you pass falsey `size` then `width`/`height` attrs are omitted
 - If you omit `className` then the `className` attr no longer has a
   trailing space (before it did)